Fix gallery album page static generation
